Revert "Allow for mapping of CMS sourced `service_api_id` to internal `serviceId`."
This reverts commit d7bd731df2bafd4175eef7e9c32d7d11828ccf59.
