+++ facilities-api/src/main/java/gov/va/api/lighthouse/facilities/api/v1/Facility.java
@@ -397,7 +397,8 @@ public final class Facility implements CanBeEmpty {
 
     /** Determine whether specified service name represents health service. */
     public static boolean isRecognizedServiceEnum(String serviceName) {
-      return isRecognizedServiceNameException(serviceName)
+      return "DentalServices".equalsIgnoreCase(serviceName)
+          || "MentalHealthCare".equalsIgnoreCase(serviceName)
           || Arrays.stream(values())
               .parallel()
               .anyMatch(hs -> hs.name().equalsIgnoreCase(serviceName));
@@ -408,15 +409,6 @@ public final class Facility implements CanBeEmpty {
       return "mentalHealthCare".equals(serviceId)
           || "dentalServices".equals(serviceId)
           || Arrays.stream(values()).parallel().anyMatch(hs -> hs.serviceId().equals(serviceId));
-    }
-
-    /**
-     * Determine whether specified service name represents known health service whose name changes
-     * between versions.
-     */
-    public static boolean isRecognizedServiceNameException(String serviceName) {
-      return "DentalServices".equalsIgnoreCase(serviceName)
-          || "MentalHealthCare".equalsIgnoreCase(serviceName);
     }
 
     @Override
