--- facilities-tests/src/test/resources/oldFormat/detailed_services_identified_by_service_api_id.json
@@ -1,80 +1,0 @@
-{
-  "detailed_services" : [
-    {
-      "service_api_id" : "covid19Vaccine",
-      "name" : "COVID-19 vaccines",
-      "active" : true,
-      "changed" : "2021-02-04T22:36:49+00:00",
-      "description_facility" : "I'm a facility service!",
-      "health_service_api_id" : "12435",
-      "appointment_leadin" : "Your VA health care team will contact you if you...more text",
-      "online_scheduling_available" : "Unknown",
-      "path" : "https://www.va.gov/bay-pines-health-care/programs/covid-19-vaccines/",
-      "appointment_phones" : [
-        {
-          "extension" : "123",
-          "label" : "Main phone changed",
-          "number" : "555-555-1212",
-          "type" : "tel"
-        },
-        {
-          "extension" : null,
-          "label" : "Main Fax",
-          "number" : "444-444-1212",
-          "type" : "fax"
-        }
-      ],
-      "referral_required" : "False",
-      "service_locations" : [
-        {
-          "additional_hours_info" : "Please use call for an apt outside...",
-          "email_contacts" : [
-            {
-              "email_address" : "georgea@va.gov",
-              "email_label" : "George Anderson"
-            },
-            {
-              "email_address" : "confirmations@va.gov",
-              "email_label" : "Confirm your appointment"
-            }
-          ],
-          "facility_service_hours" : {
-            "Friday" : "830AM-430PM",
-            "Monday" : "830AM-700PM",
-            "Sunday" : "Closed",
-            "Tuesday" : "830AM-700PM",
-            "Saturday" : "Closed",
-            "Thursday" : "830AM-600PM",
-            "Wednesday" : "ANY STRING b"
-          },
-          "appointment_phones" : [
-            {
-              "extension" : "123",
-              "label" : "Appointment phone",
-              "number" : "555-555-1212",
-              "type" : "tel"
-            },
-            {
-              "extension" : null,
-              "label" : "TTY",
-              "number" : "222-222-1212",
-              "type" : "tty"
-            }
-          ],
-          "service_location_address" : {
-            "address_line1" : "122 Main St.",
-            "address_line2" : null,
-            "building_name_number" : "Baxter Bulding",
-            "clinic_name" : "Baxter Clinic",
-            "country_code" : "US",
-            "city" : "Rochester",
-            "state" : "NY",
-            "zip_code" : "14623-1345",
-            "wing_floor_or_room_number" : "Wing East"
-          }
-        }
-      ],
-      "walk_ins_accepted" : "True"
-    }
-  ]
-}
