+++ facilities/src/main/java/gov/va/api/lighthouse/facilities/CmsOverlayControllerV0.java
@@ -205,7 +205,6 @@ public class CmsOverlayControllerV0 extends BaseCmsOverlayController {
                       // Update benefits facility services
                       facilityBenefitsServices.add(
                           Service.<DatamartFacility.BenefitsService>builder()
-                              .name(service.serviceInfo().name())
                               .serviceType(benefitsService.get())
                               .source(Source.CMS)
                               .build());
@@ -216,7 +215,6 @@ public class CmsOverlayControllerV0 extends BaseCmsOverlayController {
                           .add(
                               DATAMART_MAPPER.writeValueAsString(
                                   Service.<BenefitsService>builder()
-                                      .name(service.serviceInfo().name())
                                       .serviceType(benefitsService.get())
                                       .source(Source.CMS)
                                       .build()));
@@ -229,7 +227,6 @@ public class CmsOverlayControllerV0 extends BaseCmsOverlayController {
                       // Update health facility services
                       facilityHealthServices.add(
                           Service.<DatamartFacility.HealthService>builder()
-                              .name(service.serviceInfo().name())
                               .serviceType(healthService.get())
                               .source(Source.CMS)
                               .build());
@@ -240,7 +237,6 @@ public class CmsOverlayControllerV0 extends BaseCmsOverlayController {
                           .add(
                               DATAMART_MAPPER.writeValueAsString(
                                   Service.<HealthService>builder()
-                                      .name(service.serviceInfo().name())
                                       .serviceType(healthService.get())
                                       .source(Source.CMS)
                                       .build()));
@@ -253,7 +249,6 @@ public class CmsOverlayControllerV0 extends BaseCmsOverlayController {
                       // Update other facility services
                       facilityOtherServices.add(
                           Service.<DatamartFacility.OtherService>builder()
-                              .name(service.serviceInfo().name())
                               .serviceType(otherService.get())
                               .source(Source.CMS)
                               .build());
@@ -264,7 +259,6 @@ public class CmsOverlayControllerV0 extends BaseCmsOverlayController {
                           .add(
                               DATAMART_MAPPER.writeValueAsString(
                                   Service.<OtherService>builder()
-                                      .name(service.serviceInfo().name())
                                       .serviceType(otherService.get())
                                       .source(Source.CMS)
                                       .build()));
