+++ facilities/src/main/java/gov/va/api/lighthouse/facilities/FacilityOverlayHelper.java
@@ -2,7 +2,6 @@ package gov.va.api.lighthouse.facilities;
 
 import static org.apache.commons.lang3.ObjectUtils.isNotEmpty;
 
-import gov.va.api.lighthouse.facilities.api.TypedService;
 import java.util.stream.Collectors;
 import lombok.NonNull;
 import lombok.experimental.UtilityClass;
@@ -21,7 +20,6 @@ public class FacilityOverlayHelper {
           .detailedServices(
               datamartFacility.attributes().detailedServices().parallelStream()
                   .filter(dds -> dds.serviceInfo() != null)
-                  .filter(dds -> !dds.serviceInfo().serviceId().equals(TypedService.INVALID_SVC_ID))
                   .collect(Collectors.toList()));
     }
     return datamartFacility;
