+++ facilities/src/test/java/gov/va/api/lighthouse/facilities/CmsOverlayControllerV0Test.java
@@ -453,7 +453,6 @@ public class CmsOverlayControllerV0Test {
                         .source(Source.CMS)
                         .build(),
                     Service.<HealthService>builder()
-                        .name(CMS_OVERLAY_SERVICE_NAME_COVID_19)
                         .serviceType(HealthService.Covid19Vaccine)
                         .source(Source.CMS)
                         .build()))
@@ -686,7 +685,6 @@ public class CmsOverlayControllerV0Test {
                         .source(Source.CMS)
                         .build(),
                     Service.<HealthService>builder()
-                        .name(CMS_OVERLAY_SERVICE_NAME_COVID_19)
                         .serviceType(HealthService.Covid19Vaccine)
                         .source(Source.CMS)
                         .build()))
