+++ facilities/src/test/java/gov/va/api/lighthouse/facilities/DatamartDetailedServiceTest.java
@@ -1,12 +1,15 @@
 package gov.va.api.lighthouse.facilities;
 
+import static org.apache.commons.lang3.StringUtils.uncapitalize;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.fail;
 
 import gov.va.api.lighthouse.facilities.DatamartFacility.BenefitsService;
 import gov.va.api.lighthouse.facilities.DatamartFacility.HealthService;
 import gov.va.api.lighthouse.facilities.DatamartFacility.OtherService;
 import gov.va.api.lighthouse.facilities.api.TypeOfService;
 import gov.va.api.lighthouse.facilities.api.TypedService;
+import java.lang.reflect.Method;
 import java.util.Arrays;
 import java.util.List;
 import lombok.NonNull;
@@ -86,6 +89,220 @@ public class DatamartDetailedServiceTest {
     }
     ;
     return count;
+  }
+
+  @Test
+  @SneakyThrows
+  void isRecognizedEnumOrCovidService() {
+    Method isRecognizedEnumOrCovidServiceMethod =
+        DatamartDetailedService.class.getDeclaredMethod(
+            "isRecognizedEnumOrCovidService", String.class);
+    isRecognizedEnumOrCovidServiceMethod.setAccessible(true);
+
+    assertThat(
+            DatamartDetailedService.builder()
+                .serviceInfo(
+                    DatamartDetailedService.ServiceInfo.builder()
+                        .serviceType(TypeOfService.Health)
+                        .serviceId("empty")
+                        .name("empty")
+                        .build())
+                .build())
+        .usingRecursiveComparison()
+        .isEqualTo(
+            DatamartDetailedService.builder()
+                .serviceInfo(
+                    DatamartDetailedService.ServiceInfo.builder()
+                        .serviceId(TypedService.INVALID_SVC_ID)
+                        .name("empty")
+                        .build()));
+
+    DatamartDetailedService emptyNameDetailedService =
+        DatamartDetailedService.builder()
+            .serviceInfo(
+                DatamartDetailedService.ServiceInfo.builder()
+                    .serviceId(HealthService.Covid19Vaccine.serviceId())
+                    .name("empty")
+                    .serviceType(TypeOfService.Health)
+                    .build())
+            .build();
+    // Benefits Service
+    Arrays.stream(gov.va.api.lighthouse.facilities.api.v0.Facility.BenefitsService.values())
+        .parallel()
+        .forEach(
+            bs -> {
+              try {
+                assertThat(
+                        isRecognizedEnumOrCovidServiceMethod.invoke(
+                            emptyNameDetailedService, bs.name()))
+                    .isEqualTo(Boolean.TRUE);
+              } catch (Throwable t) {
+                fail(t.getMessage(), t);
+              }
+            });
+    Arrays.stream(DatamartFacility.BenefitsService.values())
+        .parallel()
+        .forEach(
+            bs -> {
+              try {
+                assertThat(
+                        isRecognizedEnumOrCovidServiceMethod.invoke(
+                            emptyNameDetailedService, bs.name()))
+                    .isEqualTo(Boolean.TRUE);
+              } catch (Throwable t) {
+                fail(t.getMessage(), t);
+              }
+            });
+    Arrays.stream(gov.va.api.lighthouse.facilities.api.v1.Facility.BenefitsService.values())
+        .parallel()
+        .forEach(
+            bs -> {
+              try {
+                assertThat(
+                        isRecognizedEnumOrCovidServiceMethod.invoke(
+                            emptyNameDetailedService, bs.name()))
+                    .isEqualTo(Boolean.TRUE);
+              } catch (Throwable t) {
+                fail(t.getMessage(), t);
+              }
+            });
+    // Health Service capitalized
+    Arrays.stream(gov.va.api.lighthouse.facilities.api.v0.Facility.HealthService.values())
+        .parallel()
+        .forEach(
+            hs -> {
+              try {
+                assertThat(
+                        isRecognizedEnumOrCovidServiceMethod.invoke(
+                            emptyNameDetailedService, hs.name()))
+                    .isEqualTo(Boolean.TRUE);
+              } catch (Throwable t) {
+                fail(t.getMessage(), t);
+              }
+            });
+    Arrays.stream(DatamartFacility.HealthService.values())
+        .parallel()
+        .forEach(
+            hs -> {
+              try {
+                assertThat(
+                        isRecognizedEnumOrCovidServiceMethod.invoke(
+                            emptyNameDetailedService, hs.name()))
+                    .isEqualTo(Boolean.TRUE);
+              } catch (Throwable t) {
+                fail(t.getMessage(), t);
+              }
+            });
+    Arrays.stream(gov.va.api.lighthouse.facilities.api.v1.Facility.HealthService.values())
+        .parallel()
+        .forEach(
+            hs -> {
+              try {
+                assertThat(
+                        isRecognizedEnumOrCovidServiceMethod.invoke(
+                            emptyNameDetailedService, hs.name()))
+                    .isEqualTo(Boolean.TRUE);
+              } catch (Throwable t) {
+                fail(t.getMessage(), t);
+              }
+            });
+    // Health Services uncapitalized
+    Arrays.stream(gov.va.api.lighthouse.facilities.api.v0.Facility.HealthService.values())
+        .parallel()
+        .forEach(
+            hs -> {
+              try {
+                assertThat(
+                        isRecognizedEnumOrCovidServiceMethod.invoke(
+                            emptyNameDetailedService, uncapitalize(hs.name())))
+                    .isEqualTo(Boolean.TRUE);
+              } catch (Throwable t) {
+                fail(t.getMessage(), t);
+              }
+            });
+    Arrays.stream(DatamartFacility.HealthService.values())
+        .parallel()
+        .forEach(
+            hs -> {
+              try {
+                assertThat(
+                        isRecognizedEnumOrCovidServiceMethod.invoke(
+                            emptyNameDetailedService, uncapitalize(hs.name())))
+                    .isEqualTo(Boolean.TRUE);
+              } catch (Throwable t) {
+                fail(t.getMessage(), t);
+              }
+            });
+    Arrays.stream(gov.va.api.lighthouse.facilities.api.v1.Facility.HealthService.values())
+        .parallel()
+        .forEach(
+            hs -> {
+              try {
+                assertThat(
+                        isRecognizedEnumOrCovidServiceMethod.invoke(
+                            emptyNameDetailedService, uncapitalize(hs.name())))
+                    .isEqualTo(Boolean.TRUE);
+              } catch (Throwable t) {
+                fail(t.getMessage(), t);
+              }
+            });
+    // Other Service
+    Arrays.stream(gov.va.api.lighthouse.facilities.api.v0.Facility.OtherService.values())
+        .parallel()
+        .forEach(
+            os -> {
+              try {
+                assertThat(
+                        isRecognizedEnumOrCovidServiceMethod.invoke(
+                            emptyNameDetailedService, os.name()))
+                    .isEqualTo(Boolean.TRUE);
+              } catch (Throwable t) {
+                fail(t.getMessage(), t);
+              }
+            });
+    Arrays.stream(DatamartFacility.OtherService.values())
+        .parallel()
+        .forEach(
+            os -> {
+              try {
+                assertThat(
+                        isRecognizedEnumOrCovidServiceMethod.invoke(
+                            emptyNameDetailedService, os.name()))
+                    .isEqualTo(Boolean.TRUE);
+              } catch (Throwable t) {
+                fail(t.getMessage(), t);
+              }
+            });
+    Arrays.stream(gov.va.api.lighthouse.facilities.api.v1.Facility.OtherService.values())
+        .parallel()
+        .forEach(
+            os -> {
+              try {
+                assertThat(
+                        isRecognizedEnumOrCovidServiceMethod.invoke(
+                            emptyNameDetailedService, os.name()))
+                    .isEqualTo(Boolean.TRUE);
+              } catch (Throwable t) {
+                fail(t.getMessage(), t);
+              }
+            });
+    // Covid-19 specific
+    assertThat(
+            isRecognizedEnumOrCovidServiceMethod.invoke(
+                emptyNameDetailedService, "COVID-19 vaccines"))
+        .isEqualTo(Boolean.TRUE);
+    assertThat(
+            isRecognizedEnumOrCovidServiceMethod.invoke(
+                emptyNameDetailedService,
+                uncapitalize(DatamartFacility.HealthService.Covid19Vaccine.name())))
+        .isEqualTo(Boolean.TRUE);
+    // Invalid service name
+    assertThat(
+            isRecognizedEnumOrCovidServiceMethod.invoke(emptyNameDetailedService, "No Such Name"))
+        .isEqualTo(Boolean.FALSE);
+    // Blank service name
+    assertThat(isRecognizedEnumOrCovidServiceMethod.invoke(emptyNameDetailedService, "   "))
+        .isEqualTo(Boolean.FALSE);
   }
 
   @Test
